+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "A-star/binary-heap-hash"]
-	path = A-star/binary-heap-hash
+	path = planner/binary-heap-hash
 	url = git@github.com:Cadynum/binary-heap-hash.js.git
