+++ shrdlite.html
@@ -1,5 +1,5 @@
 <!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
-<html> 
+<html>
 <head>
   <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
   <title>Shrdlite</title>
@@ -9,6 +9,7 @@
   <script type="text/javascript" src="grammar.js"></script>
   <script type="text/javascript" src="Interpreter.js"></script>
   <script type="text/javascript" src="Planner.js"></script>
+  <script type="text/javascript" src="planner/bundle.js"></script>
   <script type="text/javascript" src="shrdlite-html.js"></script>
   <!-- <script type="text/javascript" src="shrdlite-ajax.js"></script> -->
 </head>
@@ -40,9 +41,9 @@
     <strong>shrdlite.html</strong> <br>
     Copyright &copy; 2013-15, Peter Ljungl&ouml;f. <br>
     This program comes with <em>absolutely no warranty</em>. <br>
-    This is free software, and you are welcome to redistribute it under certain conditions; 
+    This is free software, and you are welcome to redistribute it under certain conditions;
     read <a href="http://www.gnu.org/licenses/gpl.html" target="__blank"><em>the license</em></a> for details.
   </div>
 </div>
-</body> 
+</body>
 </html>
